backend/requirements.txt
fastapi
uvicorn
pydantic
python-jose
sqlalchemy
psycopg2-binary